updated modal example
